third_party/blink/web_tests/fast/events/touch/touch-handler-count-expected.txt

This test checks that we correctly update the touch event handler count as event handlers are added and removed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test addEventListener/removeEventListener on the document.
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 3
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
Test setting touch handlers on the document.
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
Test addEventListener/removeEventListener on the window.
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 3
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
Test setting touch handlers on the window.
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
Test addEventListener/removeEventListener on a div.
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 3
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
Test setting touch handlers on a div.
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
Test redudant addEventListener on a div.
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
Test addEventListener/removeEventListener on a new div.
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
Test setting touch handlers on a new div.
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchEndOrCancelEventHandlerCount(document) is 0
Test that nested Documents' touch handlers are properly tracked in their parent Document.
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(nestedDocument) is 2
PASS internals.touchEndOrCancelEventHandlerCount(nestedDocument) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(nestedDocument) is 1
PASS internals.touchEndOrCancelEventHandlerCount(nestedDocument) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(nestedDocument) is 2
PASS internals.touchEndOrCancelEventHandlerCount(nestedDocument) is 2
PASS internals.touchStartOrMoveEventHandlerCount(document) is 2
PASS internals.touchEndOrCancelEventHandlerCount(document) is 2
PASS internals.touchStartOrMoveEventHandlerCount(nestedDocument) is 1
PASS internals.touchEndOrCancelEventHandlerCount(nestedDocument) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 1
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS internals.touchStartOrMoveEventHandlerCount(document) is 0
PASS internals.touchEndOrCancelEventHandlerCount(document) is 1
PASS successfullyParsed is true

TEST COMPLETE
